# Showing that random-fu is x4 faster than mwc and thus monad-bayes