--- a/.envrc
+++ b/.envrc
@@ -1,2 +1,3 @@
+use nix
 # Make sure you have direnv >= 2.30
-use flake --extra-experimental-features nix-command --extra-experimental-features flakes
+# use flake --extra-experimental-features nix-command --extra-experimental-features flakes