```
ghc samplePerformance.hs 
[1 of 1] Compiling Main             ( samplePerformance.hs, samplePerformance.o )
Linking samplePerformance ...

./samplePerformance 
benchmarking dists/StdGen/stdNormal/single sample
time                 45.20 ns   (45.16 ns .. 45.25 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 45.28 ns   (45.22 ns .. 45.41 ns)
std dev              304.5 ps   (173.5 ps .. 545.5 ps)

benchmarking dists/StdGen/stdNormal/single sample MWC
time                 157.0 ns   (156.9 ns .. 157.2 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 157.0 ns   (156.9 ns .. 157.3 ns)
std dev              600.7 ps   (338.0 ps .. 990.2 ps)

benchmarking dists/StdGen/stdNormal/single sample monad bayes
time                 157.2 ns   (156.9 ns .. 157.6 ns)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 158.5 ns   (157.1 ns .. 163.6 ns)
std dev              8.239 ns   (1.448 ns .. 17.33 ns)
variance introduced by outliers: 71% (severely inflated)
```