@turion 2023-10-13
> I'd also like to benchmark sampling in Haskell against other languages. But is a PR here the best place to do these experiments?

Yes, I guess it is, but in a separate PR (and a separate benchmarking suite).