@idontgetoutmuch 2023-10-13
> We may be comparing apples and oranges here and one thing I plan to try is replacing `mwc-random` by `random-fu`.

@turion I just tried

```
newtype SamplerU g m a = SamplerU {runSamplerU :: ReaderT g m a} deriving (Functor, Applicative, Monad, MonadIO)

instance (StatefulGen g m, MonadReader g m) => MonadDistribution (SamplerU g m) where
  random = SamplerU undefined

  uniform a b = SamplerU undefined
  normal m s = undefined -- SamplerU (ReaderT $ (RF.sample (RF.normal m s)))
  gamma shape scale = SamplerU undefined
  beta a b = SamplerU undefined

  bernoulli p = undefined
  categorical ps = undefined
  geometric p = undefined
```

but have type errors (if I uncomment the definition for `normal`). I have to give up for the weekend as we have visitors but you had some time, maybe you could take a butcher's.