@idontgetoutmuch 2023-10-14
Replaced by https://github.com/tweag/monad-bayes/pull/323